+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ## [Unreleased]
 
-### Added
-
+-   Support for kedro-mlflow>=0.7
 -   Support of Google Vertex AI Pipelines (EXPERIMENTAL)
 
 ## [0.3.1] - 2021-05-25